sdk/python/requirements.txt
requests==2.31.0
python-socketio[client]==5.11.1
websocket-client==1.7.0
base58==2.1.1
PyNaCl==1.5.0
